move virtualized rendering of TimingChart to Render.Components.TimingView
